test(sync): split 5 oversized test files (sync-vmv4)
Split conflict-resolution (1548), schemas (1087), ws-client-adapter
(1043), document-lifecycle (909), and sync-engine-runtime-hardening
(769) test files into 18 smaller test files (each ≤500 LOC) plus 5
shared helper modules. All 139 source tests preserved; full sync suite
(951 tests) passes.

DRY pass extracts sodium-session setup, schema doc factories,
MockWebSocket/TestHarness, and SyncEngine bootstrap helpers to per-
domain fixtures under __tests__/helpers/.
